requirements.txt
fastapi
uvicorn
pydantic
supabase
python-dotenv
requests
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
joblib
geopy